requirements.txt
fastapi
uvicorn

psycopg2-binary
sqlalchemy
python-dotenv

pydantic[email]
passlib[bcrypt]

pytest
requests

python-jose

pytest-cov

python-multipart